requirements.txt
hikari 
hikari-lightbulb
python-dotenv
openai